refactoring: Extract all the kube-aws specific text/json/userdata templating to generate files
from the config package to their own packages so that we can be more explicit about it's existence

This also allow reusing in the upcoming node pools feature.
